test(flat-table): typescript refactor
